refactor: rename component files for easier import
