+++ src/components/app/index.tsx
@@ -1,4 +1,4 @@
-import MainPage from '../../pages/main-page/main-page';
+import MainPage from '../../pages/main-page';
 
 function App(): JSX.Element {
   return(
