+++ src/components/place-card/index.tsx
