+++ src/pages/favorites-page/index.tsx
