+++ src/pages/offer-page/index.tsx
